correção ASPP
