--- src/SME.SGP.Aplicacao/Interfaces/CasosDeUso/AtribuicaoDeResponsaveis/IRemoverAtribuicaoResponsaveisCoreSSO.cs
@@ -1,13 +1,0 @@
-﻿using SME.SGP.Infra;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace SME.SGP.Aplicacao.Interfaces
-{ 
-    public interface IRemoverAtribuicaoResponsaveisCoreSSO :  IUseCase<MensagemRabbit, bool>
-    {
-    }
-}
