+++ src/SME.SGP.IoC/RegistraDependencias.cs
@@ -1308,7 +1308,6 @@ namespace SME.SGP.IoC
             services.TryAddScoped<IRemoverAtribuicaoResponsaveisSupervisorPorDreUseCase, RemoverAtribuicaoResponsaveisSupervisorPorDreUseCase>();
             services.TryAddScoped<IRemoverAtribuicaoResponsaveisPAAIPorDreUseCase, RemoverAtribuicaoResponsaveisPAAIPorDreUseCase>();
             services.TryAddScoped<IRemoverAtribuicaoResponsaveisASPPPorDreUseCase, RemoverAtribuicaoResponsaveisASPPPorDreUseCase>();
-            services.TryAddScoped<IRemoverAtribuicaoResponsaveisCoreSSO, RemoverAtribuicaoResponsaveisCoreSSO>();
 
 
         }
